+++ crucible/CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add new helpers for extracting `SymGlobalState`s: `exec{Result,State}Globals`.
 * Add `typedOverride` for constructing `TypedOverride`s with statically-known
   signatures.
+* Add `bindTypedOverride` for binding `TypedOverride`s to `FnHandle`s.
 
 # 0.7.2 -- 2025-03-21